消除黄色警告

+++ EduSohoAndroidClient/edusoho_v3/src/main/java/com/edusoho/kuozhi/clean/bean/OrderInfo.java
@@ -17,8 +17,8 @@ public class OrderInfo implements Serializable {
     public String coinName;
     public float cashRate;
     public String priceType;
-    public int coinPayAmount;
-    public int maxCoin;
+    public float coinPayAmount;
+    public float maxCoin;
     public String duration;
     public int fullCoinPayable;
     public String title;
